pyproject.toml
[tool.poetry]
name = "novelty-experiment"
version = "0.1.0"
description = ""
authors = ["aakarsh <aakarsh@poocho.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
duckdb = "^1.2.1"
numpy = "^2.2.3"
scikit-learn = "^1.6.1"
openai = "^1.65.5"
setuptools = "^76.0.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
